chatbot/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.6
qdrant-client==1.7.0
psycopg[binary]==3.1.13
asyncpg==0.29.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
tiktoken==0.5.2
markdown==3.5.1
python-multipart==0.0.6
requests
sentence-transformers>=2.2.2
transformers>=4.36.0
torch>=2.1.0
huggingface-hub>=0.20.0
